simplify chrome api
